Update styles so NavBar doesn't overlap <main>
To make the Hero image height be the remainder of
the screen, I set min-height: calc(100vh - 72px),
where 72px is the height of the NavBar. Sometimes
when I refresh the page, there is initially a
small gap between the end of the image and the end
of the viewport, but then it fills it. I don't
know if this is because I set the min-height in
the style property, but should consider creating a
new tailwind class to see if it solves the issue
of there sometimes being that gap.