Update Car interface with Make & Model interfaces
The Make and Model properties of Car interface
were strings. Now their types are the Make and
Model interfaces. This is so when the user clicks
the Request Import button on a vehicle page, I
will be able to take the user to step_4 of the
request-import-form and autofill the first three
steps with the vehicle's information