Refactor & update styles of Selector
Wanted to widen ProductionYearsSelector because
it felt like a waste of horizontal space, which
resulted in unnecessarily used vertical space.

Added two divs inside li elements and used those
for the gray background color. Before, when the
grid was stretched out, the rounded li elements
would start to look bad. This fixed that problem.