Remove body height so pages' min-height-svh works
When body { height: 100% } was in place,
min-height wouldn't work on layout pages. It cut
off content that overflowed the body. Removing
body { height: 100% } enables child pages to
effectively set a min-height, where the page fits
properly on mobile and desktop when the height of
its children doesn't exceed it, and when the
height of its children does exceed it, it can grow
with it.