Redirect user home on ContactForm if no userData
This is for if the user presses browser back button on /post-request page. As of now, if I send user to first page of form, the make option is still selected. This can probably be fixed, but this is my solution for now. I will need to come back to this because it seems like it can be better.