Update styles so button & table aren't md:w-full
Put button and table in same grid as ImageGrid,
and keep width the same as the first image after
md: viewport width.

The divs that the button and table are now in
cause them to still be visible when
ImageGalleryModal is open, so I will set a higher
z-index for ImageGalleryModal to z-30.